webapi/webapi-contactsmanager-sysapps-tests/contactsmanager/ContactName_attributes.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xu, Kang <kangx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>ContactsManager Test: ContactName_attributes</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="http://www.w3.org/2012/sysapps/contacts-manager-api/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>
var contactName;

setup(function () {
  contactName = new ContactName({
    displayName: 'display',
    honorificPrefixes: ['prefixes'],
    givenNames: ['John'],
    additionalNames: ['additional'],
    familyNames: ['Doe'],
    honorificSuffixes: ['suffixes'],
    nicknames: ['nick']
  });
});

test(function () {
  assert_true("displayName" in contactName, "The ContactName.displayName exists");
  assert_equals(typeof contactName.displayName, "string", "The type of ContactName.displayName");
}, "Check if the attribute ContactName.displayName exists and typeof string");

test(function () {
  assert_true("honorificPrefixes" in contactName, "The ContactName.honorificPrefixes exists");
  assert_equals(typeof contactName.honorificPrefixes, "object", "The type of ContactName.honorificPrefixes");
}, "Check if the attribute ContactName.honorificPrefixes exists and typeof object");

test(function () {
  assert_true("givenNames" in contactName, "The ContactName.givenNames exists");
  assert_equals(typeof contactName.givenNames, "object", "The type of ContactName.givenNames");
}, "Check if the attribute ContactName.givenNames exists and typeof object");

test(function () {
  assert_true("additionalNames" in contactName, "The ContactName.additionalNames exists");
  assert_equals(typeof contactName.additionalNames, "object", "The type of ContactName.additionalNames");
}, "Check if the attribute ContactName.additionalNames exists and typeof object");

test(function () {
  assert_true("familyNames" in contactName, "The ContactName.familyNames exists");
  assert_equals(typeof contactName.familyNames, "object", "The type of familyNames.displayName");
}, "Check if the attribute ContactName.familyNames exists and typeof object");

test(function () {
  assert_true("honorificSuffixes" in contactName, "The ContactName.honorificSuffixes exists");
  assert_equals(typeof contactName.honorificSuffixes, "object", "The type of ContactName.honorificSuffixes");
}, "Check if the attribute ContactName.honorificSuffixes exists and typeof object");

test(function () {
  assert_true("nicknames" in contactName, "The ContactName.nicknames exists");
  assert_equals(typeof contactName.nicknames, "object", "The type of ContactName.nicknames");
}, "Check if the attribute ContactName.nicknames exists and typeof object");
</script>
